Basic playing around with stuff. This will connect and it can vaguely interact with discord. unfortunately it seems the mudae bot doesn't obey it's commands.
TODO:
See if Mudae will react to this bot reacting to emojis. So for instance paging through an ima.

With that and an auto hotkey tool to type into discord, it shouldn't be too hard to still scrape the data, it will just be slower so instead of like a daily update it might be like a weekly update.