Massive Overhaul, haven't committed in a while
Worked out an AutoHotKey script that will read from the Commands.txt, the last line in it, assuming it's different. It can be commanded to click x number of times or to type and enter a certain command. This is my interface with interacting with the mudae bot through discord, as the mudae bot ignores all actions from bots.

I decomposed my scraper class into multiple classes with my main class being the discord interface where i can handle on_message and on_message_edit. It hands off the message objects to a mudae class which then parses, scrapes, manages the whole thing.

The mudae class creates a state object which will maintain the bundles and series it needs to iterate through with a queue, but for now it just has some basic objects like the current filename, the last ran command, etc...

I have some static helper classes for things like giving commands to ahk, writing out to our files the scraped data, and for parsing various things. I've probably gone a bit too far into the whole decomposing into objects but ohwell.

This current edition will parse and get all the characters and series from Trigger (my test seed bundle). As far as I can tell it works exactly as desired when given a seed bundle.

The next major step is to work through the building of our Bundle Queue. Meaning we need to take that seed bundle, get all the series associated with it, get all the bundles for all of those series, make sure we haven't ran them before (this will already take forever as it is), and basically build up this queue of bundles we need to scrape. Then just start powering through scraping all of them.